// src/pages/card-indexes/interfaces/card-index.interface.ts
export interface CardIndexInterface {
  _id: string;
  code: string;
  verification: string;
  consumer: string;
  zone2: string;
  track: string;
  id: string;
  international: string;
  start: string;
  end: string;
  signal_level: string;
  order_number_for_stream: string;
  creatorId?: string;
  created_at?: string;
  updated_at?: string;
  __v?: number;
}